Added example test data.
Removed example_populate.py and reformatted the entry of 35 users and 50 random cases to populate.py. This can be generated with the command run_foreman.py setup_example <config>.

Generation of QR codes is now done after evidence object is created as it relies on having the evidence object ID as the QR png name, which it doesn't know until committed.

Empty folder "files" created for the setup_example and also if user's have a text file to generate the next case name (not implemented yet).