Task Types & Categories can now be added/removed
In the admin panel Task Types and Categories can be added and removed. When a task type is removed and there are Tasks with that type; their task type is  changed to "Undefined" which is a hidden task type. When the user edits the task they can change to a new type, but by having the "Undefined" as a task type, it means the system does not break when the task types are removed.

Tak Categories can oly be removed once all task types have been moved out or deleted.