Uploading of case & task name files
Added code to allow admin to upload a text file for next bunch of case and/or task names. Foreman alerts admin is they have run out of names and then defaults to a numeric increment for the next name. Also new alerts admin if they haven't changed their default password and a 'hint' when they start from scratch that they want to add users.

Also got the default population code set up ready for the unit tests, so creates a standard database for testing.