Uploading of files to tasks
Added ability to upload files to the task. The user must give a title, comment and file. The file mete data and hash are stored in the database.

Also added some error checking to see if the installation has the to PDf and RTF libraries. Shows the user an error when trying to export to these if the libraries are not present.