Added case priority
Cases can now be assigned a priority. Admin's can create new priorities with an assigned colour, and one as the default. They can also remove priorities (but not the last one - one priority must be left and this just becomes the default for everything). The colour coded priority is available in case & task tables along the left hand side and in the case/task views. The default set for Foreman is low, normal, high and critical.