Timesheets more managable
Previously the timesheet pulled every task/case since the dawn of time for the user. Now it is a week by week basis, with tasks and cases allowing for timesheets between their creation date and date closed/archived. Therefore users are allowed to enter times before they are assigned things (i.e. something was created but only worked upon officially a week later when the user was assigned) - in real life not everything is assigned in a timely manner, and investigators may go on calls / do prep work before being assigned, so this way, they can enter in hours from the moment the case/task was entered onto the system.

Commented out urls in the JQuery CSS to images that don't exist.

Managers of users can now see the users active cases/tassk on their profile page.