Evidence retention policy
The admin can set a retention policy so that after a set number of moths after the evidence was set to archived, an email will be sent to the admins and custodian of the evidence that the evidence can now be destroyed. This update also includes 4 evidence statuses: inactive, active, archived and destroyed. Archived and destroyed are essentially the same, allow no editing, but may be useful to differenciate between evidence in storage and evidence that has been destroyed. To set up the automatic emailing for the destruction of evidence, a CRON job or task scheduler must be set up with a daily job to run_foreman.py scheduled_tasks <link to config file>