Added new email alert
Case managers can now get an email if a requester adds a new task to their case.

Bug fixes and improvements: Removed some unused imports and fixed a bug that showed a link to allow requesters to assign themselves as case managers. Changed the Foreman runserver argument to have an underscore like the rest of the arguments setup_example, run_tests etc